# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SMART CLINIC - Healthcare SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics/{clinicId}/patients/{patientId}/lab-requests:
    get:
      operationId: listLabRequests
      tags: [patients]
      summary: List lab requests for a patient
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: patientId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of lab requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabRequest"
    post:
      operationId: createLabRequest
      tags: [patients]
      summary: Create a lab request (doctor/admin only)
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: patientId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabRequestInput"
      responses:
        "201":
          description: Created lab request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabRequest"
  /clinics/{clinicId}/patients/{patientId}/lab-requests/{requestId}:
    patch:
      operationId: updateLabRequest
      tags: [patients]
      summary: Update lab request status
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: patientId
          required: true
          schema:
            type: string
        - in: path
          name: requestId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabRequestUpdate"
      responses:
        "200":
          description: Updated lab request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabRequest"
  /clinics/{clinicId}/patients/{patientId}/lab-requests/{requestId}/results:
    get:
      operationId: listLabResults
      tags: [patients]
      summary: Get lab results for a request
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: patientId
          required: true
          schema:
            type: string
        - in: path
          name: requestId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lab results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabResult"
    post:
      operationId: submitLabResult
      tags: [patients]
      summary: Submit a lab result (lab technician/admin only)
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: patientId
          required: true
          schema:
            type: string
        - in: path
          name: requestId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabResultInput"
      responses:
        "201":
          description: Created lab result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabResult"

  # ===== PHARMACY QUEUE =====
...
  /clinics/{clinicId}/lab/queue:
    get:
      operationId: getLabQueue
      tags: [lab]
      summary: Get lab request queue for clinic
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Lab queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabRequest"
...
components:
  schemas:
...
    LabRequestInput:
      type: object
      required: [testName]
      properties:
        testName:
          type: string
        testCategory:
          type: string
          enum: [blood, urine, stool, imaging, swab, other]
        urgency:
          type: string
          enum: [routine, urgent, stat]
        notes:
          type: ["string", "null"]
    LabRequestUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
    LabRequest:
      type: object
      required: [id, clinicId, patientId, doctorId, requestCode, testName, testCategory, urgency, status, createdAt]
      properties:
        id:
          type: string
        clinicId:
          type: string
        patientId:
          type: string
        patientName:
          type: ["string", "null"]
        patientCode:
          type: ["string", "null"]
        doctorId:
          type: string
        doctorName:
          type: ["string", "null"]
        requestCode:
          type: string
        testName:
          type: string
        testCategory:
          type: string
        urgency:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        createdAt:
          type: string
          format: date-time
    LabResultInput:
      type: object
      required: [resultText]
      properties:
        resultText:
          type: string
        resultNotes:
          type: ["string", "null"]
        status:
          type: string
          enum: [preliminary, final, amended]
    LabResult:
      type: object
      required: [id, clinicId, labRequestId, patientId, resultText, status, createdAt]
      properties:
        id:
          type: string
        clinicId:
          type: string
        labRequestId:
          type: string
        patientId:
          type: string
        technicianId:
          type: ["string", "null"]
        technicianName:
          type: ["string", "null"]
        resultText:
          type: string
        resultNotes:
          type: ["string", "null"]
        status:
          type: string
          enum: [preliminary, final, amended]
        createdAt:
          type: string
          format: date-time

    # ===== PHARMACY SCHEMAS =====
...